# specs/001-bloom-tutor-app/contracts/api.yaml
openapi: 3.0.3
info:
  title: Bloom GCSE Mathematics Tutor API
  description: |
    FastAPI backend for Bloom tutoring application.
    Uses htmx for frontend interactions - most endpoints return HTML fragments.
  version: 1.0.0
  contact:
    name: Bloom Development Team
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /session/start:
    post:
      tags: [session]
      summary: Start new tutoring session
      description: |
        Creates a new session for specified subtopic.
        If active session exists for different subtopic, marks it as abandoned.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [subtopic_id]
              properties:
                subtopic_id:
                  type: integer
                  example: 101
      responses:
        '200':
          description: HTML chat interface
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Invalid subtopic ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /session/resume:
    post:
      tags: [session]
      summary: Resume last active session
      description: Restores conversation state from agent checkpoint
      responses:
        '200':
          description: HTML chat interface with restored history
          content:
            text/html:
              schema:
                type: string
        '404':
          description: No active session found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /chat:
    get:
      tags: [student]
      summary: Chat interface page
      description: |
        Full-page chat UI for current session. Includes "+ New Chat" button in header to allow students to end the current session and return to syllabus (FR-021)
      responses:
        '200':
          description: HTML chat page
          content:
            text/html:
              schema:
                type: string
        '404':
          description: No active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /chat/messages:
    get:
      tags: [student]
      summary: Get chat message history
      description: |
        Returns messages for current session.
        Used by htmx polling: `hx-get="/chat/messages" hx-trigger="every 2s"`
      parameters:
        - name: session_id
          in: query
          required: false
          schema:
            type: integer
          description: Session ID (if not active session)
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Only return messages after this timestamp
      responses:
        '200':
          description: HTML fragment with message elements
          content:
            text/html:
              schema:
                type: string
                example: |
                  <div class="message tutor">
                    <p>Let's work on fractions!</p>
                    <span class="timestamp">14:32</span>
                  </div>
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
  /chat/message:
    post:
      tags: [student]
      summary: Send student message
      description: |
        Student submits answer or question.
        Triggers LangGraph agent to evaluate and respond.
        Returns tutor's response + updated UI state (calculator visibility, etc.)
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  example: "1.15"
      responses:
        '200':
          description: HTML fragment with tutor response and UI updates
          content:
            text/html:
              schema:
                type: string
        '500':
          description: LLM API failure (friendly error message per FR-018)
          content:
            text/html:
              schema:
                type: string
                example: |
                  <div class="alert error">
                    I'm having trouble connecting. Please try again.
                    <button hx-post="/chat/retry">Retry</button>
                  </div>
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
          example: 42
        role:
          type: string
          enum: [student, tutor]
          example: "tutor"
        content:
          type: string
          example: "Great job! Let's try another one."
        timestamp:
          type: string
          format: date-time
          example: "2025-11-24T14:32:15"
    Error:
      type: object
      properties:
        error:
          type: string
          example: "No active session found"
        code:
          type: string
          example: "SESSION_NOT_FOUND"
...